LayoutTests/fast/forms/date/date-editable-components/date-picker-update-on-edit-expected.txt

Tests that a user edit in the control updates the date picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that picker is displayed
PASS showingPicker is true

Check initial value
PASS input.value is "2020-09-16"
PASS input.valueAsNumber is pickerValue

Check picker value after user edit
PASS input.value is "2020-03-24"
PASS input.valueAsNumber is pickerValue

Check picker value after programmatic change
PASS input.value is "2020-12-30"
PASS input.valueAsNumber is not pickerValue

PASS successfullyParsed is true

TEST COMPLETE
